Created runArchTests function
